refactor: move Validation module and work views to view.py
