+++ src/ramstk/views/gtk3/books/modulebook.py
@@ -19,8 +19,7 @@ from ramstk.views.gtk3.function import FunctionModuleView
 # from ramstk.views.gtk3.hardware import HardwareModuleView
 from ramstk.views.gtk3.requirement import RequirementModuleView
 from ramstk.views.gtk3.revision import RevisionModuleView
-
-# from ramstk.views.gtk3.validation import ValidationModuleView
+from ramstk.views.gtk3.validation import ValidationModuleView
 from ramstk.views.gtk3.widgets import RAMSTKBaseBook
 
 
@@ -51,7 +50,7 @@ class RAMSTKModuleBook(RAMSTKBaseBook):
             "function": FunctionModuleView(configuration, logger),
             "requirement": RequirementModuleView(configuration, logger),
             # "hardware": HardwareModuleView(configuration, logger),
-            # "validation": ValidationModuleView(configuration, logger),
+            "validation": ValidationModuleView(configuration, logger),
         }
 
         # Initialize private list attributes.
