+++ src/ramstk/views/gtk3/validation/view.pyi
@@ -1,0 +1,51 @@
+# Standard Library Imports
+from typing import Any, Dict
+
+# RAMSTK Package Imports
+from ramstk.configuration import RAMSTKUserConfiguration as RAMSTKUserConfiguration
+from ramstk.logger import RAMSTKLogManager as RAMSTKLogManager
+from ramstk.views.gtk3 import Gtk as Gtk
+from ramstk.views.gtk3 import _ as _
+from ramstk.views.gtk3.widgets import RAMSTKMessageDialog as RAMSTKMessageDialog
+from ramstk.views.gtk3.widgets import RAMSTKModuleView as RAMSTKModuleView
+from ramstk.views.gtk3.widgets import RAMSTKPanel as RAMSTKPanel
+from ramstk.views.gtk3.widgets import RAMSTKWorkView as RAMSTKWorkView
+
+# RAMSTK Local Imports
+from . import ValidationTaskDescriptionPanel as ValidationTaskDescriptionPanel
+from . import ValidationTaskEffortPanel as ValidationTaskEffortPanel
+from . import ValidationTreePanel as ValidationTreePanel
+
+class ValidationModuleView(RAMSTKModuleView):
+    _module: str
+    _tablabel: str
+    _tabtooltip: str
+    _lst_mnu_labels: Any
+    _lst_tooltips: Any
+    _pnlPanel: Any
+    def __init__(
+        self, configuration: RAMSTKUserConfiguration, logger: RAMSTKLogManager
+    ) -> None: ...
+    def do_request_delete(self, __button: Gtk.ToolButton) -> None: ...
+    _record_id: Any
+    def _do_set_record_id(self, attributes: Dict[str, Any]) -> None: ...
+    def __make_ui(self) -> None: ...
+
+class ValidationGeneralDataView(RAMSTKWorkView):
+    _module: str
+    _tablabel: str
+    _tabtooltip: str
+    _lst_callbacks: Any
+    _lst_icons: Any
+    _lst_mnu_labels: Any
+    _lst_tooltips: Any
+    _pnlTaskDescription: Any
+    _pnlTaskEffort: Any
+    def __init__(
+        self, configuration: RAMSTKUserConfiguration, logger: RAMSTKLogManager
+    ) -> None: ...
+    def _do_request_calculate(self, __button: Gtk.ToolButton) -> None: ...
+    def _do_request_calculate_all(self, __button: Gtk.ToolButton) -> None: ...
+    _record_id: Any
+    def _do_set_record_id(self, attributes: Dict[str, Any]) -> None: ...
+    def __make_ui(self) -> None: ...
